+++ R/catalog_apply.r
@@ -166,8 +166,9 @@ catalog_apply <- function(ctg, func, func_args = NULL, ...)
   res       <- 1
   progress  <- progress(ctg)
   ncores    <- cores(ctg)
+  stopearly <- stop_early(ctg)
   clusters  <- catalog_makecluster(ctg, res)
-  output    <- cluster_apply(clusters, cluster_apply_func, ncores, progress, func = func, ctg = ctg, func_args = func_args, ...)
+  output    <- cluster_apply(clusters, cluster_apply_func, ncores, progress, stopearly, func = func, ctg = ctg, func_args = func_args, ...)
   return(output)
 }